requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pillow==10.4.0
torch==2.9.1
torchvision==0.24.1
requests==2.32.3
numpy==1.26.4
sqlalchemy==2.0.36